requirements.txt
python-jose[cryptography]
python-multipart
pathlib
jinja2
pytest
argon2-cffi
email-validator
python-dotenv
cassandra-driver
uvicorn
fastapi